requirements.txt
streamlit==1.44.0
openai-agents==0.0.7
pydantic==2.10.6
yfinance==0.2.55
python-dotenv==1.1.0
asyncio
uvicorn==0.34.0
fastapi==0.115.12
rich
pydantic-settings
